[Chore] Seperate enum script and move file directory
