+++ ItaCH_Smash_Legends/Assets/Script/UI/IPanel.cs
