+++ ItaCH_Smash_Legends/Assets/Script/UI/LegendSelectUI.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
-using Util;
+using Util.Enum;
 
 public class LegendSelectUI : MonoBehaviour
 {
