+++ ItaCH_Smash_Legends/Assets/Script/UI/LegendSelectUI.cs.meta
