+++ ItaCH_Smash_Legends/Assets/Script/UI/LegendUI.cs
