+++ ItaCH_Smash_Legends/Assets/Script/UI/RemovePanelButton.cs
