+++ ItaCH_Smash_Legends/Assets/Script/UI/ScoreSet.cs
