+++ ItaCH_Smash_Legends/Assets/Script/UI/SettingPanel.cs
