+++ ItaCH_Smash_Legends/Assets/Script/Util/Path.cs
@@ -1,0 +1,28 @@
+using System.Collections;
+using System.Collections.Generic;
+using System.Text;
+using UnityEngine;
+using Util.Enum;
+
+namespace Util.Path
+{
+    public static class FilePath
+    {
+        #region 파일 경로
+        public const string LegendMenuUIPath = "UI/LegendMenu";
+        public const string LegendMenuButtonPath = "UI/LegendMenuButton";
+        public const string SettingUIPath = "UI/SettingPanel";
+        public const string SettingButtonPath = "UI/SettingButton";
+        public const string MatchingUIPath = "UI/MatchingPanel";
+        public const string MatchingButtonPath = "UI/GameStartButton";
+        #endregion
+        public static string GetLobbyCharacterPath(CharacterType characterType)
+        {
+            StringBuilder stringBuilder = new StringBuilder();
+            string characterName = characterType.ToString();
+            stringBuilder.Append("Map/LobbyWorld/");
+            stringBuilder.Append(characterName);
+            return stringBuilder.ToString();
+        }
+    }
+}
