+++ ItaCH_Smash_Legends/Assets/Script/Util/Path.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1bdee1a669620ca4e810ebf336bfec48
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
